Update directory structure for tests.
