+++ Tests/HelperExtensions.swift
@@ -4,6 +4,7 @@
 //
 
 import Foundation
+import Moya
 
 // MARK: - Provider support
 extension String {
@@ -14,4 +15,8 @@ extension String {
   var data: NSData {
     return self.dataUsingEncoding(NSUTF8StringEncoding)!
   }
+}
+
+func url(route: TargetType) -> String {
+  return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
 }
